+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Add "Copy actor ID" menu item to profile menu (only for admins).
+
 ## [2.6.0] - 2023-12-25
 
 ### Added