+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add "Copy actor ID" menu item to profile menu (only for admins).
+- Display reply indicator when replying to one's own post.
 
 ## [2.6.0] - 2023-12-25