+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add loader to tag timeline.
 - Change representation of an attachment of unknown type.
+- Allow interactions with images in Gallery.
 
 ## [2.6.0] - 2023-12-25