+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [2.7.0] - 2024-01-12
+
 ### Added
 
 - Add "Copy actor ID" menu item to profile menu (only for admins).